String Guide
A simple HTML interface to help Guitarist make the best selection of string size for each tuning.

There are tons of songs they need a specific tuning to be well played. So this little App helps you decide which one is the best. It includes a simple link to amazon.

The app is based on the following video.

https://www.youtube.com/watch?v=UkG8Ga9aFfo

I recommend Earni Ball string because the are the widely common know brand, but you can choose anyone, what really matters is the size of the string.

I hope you enjoy it.